feat - move class,  Currency
