+++ src/main/java/co/invest72/exchange_rate/domain/service/Bank.java
@@ -3,7 +3,7 @@ package co.invest72.exchange_rate.domain.service;
 import java.math.BigDecimal;
 
 import co.invest72.exchange_rate.domain.ExchangeRateProvider;
-import co.invest72.money.domain.Currency;
+import co.invest72.exchange_rate.domain.Currency;
 import co.invest72.money.domain.Money;
 
 public class Bank {
