+++ src/main/java/co/invest72/investment/domain/investment/CashInvestment.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 import co.invest72.investment.domain.Investment;
 import co.invest72.investment.domain.InvestmentAmount;
 import co.invest72.investment.domain.tax.TaxType;
-import co.invest72.money.domain.Currency;
+import co.invest72.exchange_rate.domain.Currency;
 import co.invest72.money.domain.Money;
 
 public class CashInvestment implements Investment {
